Move grpc into api
