+++ app/server/main.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"TUM-Live/api"
-	"TUM-Live/api_grpc"
 	"TUM-Live/dao"
 	"TUM-Live/model"
 	"TUM-Live/tools"
@@ -171,7 +170,7 @@ func initCron() {
 	//Flush stale sentry exceptions and transactions every 5 minutes
 	_, _ = cronService.AddFunc("0-59/5 * * * *", func() { sentry.Flush(time.Minute * 2) })
 	//Look for due streams and notify workers about them
-	_, _ = cronService.AddFunc("0-59 * * * *", api_grpc.NotifyWorkers)
+	_, _ = cronService.AddFunc("0-59 * * * *", api.NotifyWorkers)
 	cronService.Start()
 }
